CLIENT UNBLOCK

+++ src/test/java/redis/clients/jedis/tests/commands/ClientCommandsTest.java
@@ -103,13 +103,13 @@ public class ClientCommandsTest extends JedisCommandTestBase {
   @Test
   public void clientUnblock() throws InterruptedException, TimeoutException {
     long clientId = client.clientId();
-    assertEquals(0, jedis.clientUnblock(clientId, UnblockType.ERROR).longValue());
+    assertEquals(0, jedis.clientUnblock(clientId, UnblockType.ERROR));
     Future<?> future = Executors.newSingleThreadExecutor().submit(() -> client.brpop(100000, "foo"));
 
     try {
       // to make true command already executed
       TimeUnit.MILLISECONDS.sleep(500);
-      assertEquals(1, jedis.clientUnblock(clientId, UnblockType.ERROR).longValue());
+      assertEquals(1, jedis.clientUnblock(clientId, UnblockType.ERROR));
       future.get(1, TimeUnit.SECONDS);
     } catch (ExecutionException e) {
       assertEquals("redis.clients.jedis.exceptions.JedisDataException: UNBLOCKED client unblocked via CLIENT UNBLOCK", e.getMessage());
